+++ changelog.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-04-10
+
+### Added
+
+- "Public mode" - now the bot can run without a "master" OpenAI API Key, and every user will have to provide their own.
+
+### Changed
+
+- Codebase has been refactored.
+
 ## [0.3.4] - 2023-04-08
 
 ### Added
@@ -85,7 +95,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Flake8 and Mypy setups.
 - GitHub Action for linters.
 
-[Unreleased]: https://github.com/s-nagaev/chibi/compare/v0.3.2...HEAD
+[Unreleased]: https://github.com/s-nagaev/chibi/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/s-nagaev/chibi/compare/v0.3.4...v0.4.0
+[0.3.4]: https://github.com/s-nagaev/chibi/compare/v0.3.3...v0.3.4
+[0.3.3]: https://github.com/s-nagaev/chibi/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/s-nagaev/chibi/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/s-nagaev/chibi/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/s-nagaev/chibi/compare/v0.2.1...v0.3.0